move js loader into own package
