+++ packages/cosmiconfig/src/defaults.ts
@@ -1,4 +1,4 @@
-import { loadJs, loadJsSync } from '@cosmiconfig/base/loaders';
+import { loadJs, loadJsSync } from '@cosmiconfig/js-loader';
 import { loadTs, loadTsSync } from '@cosmiconfig/typescript-loader';
 import { loadYaml } from '@cosmiconfig/yaml-loader';
 import { loadJson } from './loaders';
