+++ packages/cosmiconfig/tsconfig.build.json
@@ -10,6 +10,7 @@
   "references": [
     { "path": "../base/tsconfig.build.json" },
     { "path": "../types/tsconfig.build.json" },
+    { "path": "../js-loader/tsconfig.build.json" },
     { "path": "../typescript-loader/tsconfig.build.json" },
     { "path": "../yaml-loader/tsconfig.build.json" }
   ]
