+++ packages/cosmiconfig/tsconfig.json
@@ -8,6 +8,7 @@
   "references": [
     { "path": "../base/tsconfig.build.json" },
     { "path": "../types/tsconfig.build.json" },
+    { "path": "../js-loader/tsconfig.build.json" },
     { "path": "../typescript-loader/tsconfig.build.json" },
     { "path": "../yaml-loader/tsconfig.build.json" }
   ]
