+++ packages/js-loader/tsconfig.build.json
@@ -6,8 +6,5 @@
     "outDir": "./dist"
   },
   "include": ["src"],
-  "references": [
-    { "path": "../types/tsconfig.build.json" },
-    { "path": "../js-loader/tsconfig.build.json" }
-  ]
+  "references": [{ "path": "../types/tsconfig.build.json" }]
 }
